+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ### Breaking
 * throw exception when calling `(-)`
 ### Features
-* `let-values` and `let*-values` syntax macros
+* add `let-values`, `let*-values` and `define-values` syntax macros
 ### Bugfix
 * fix empty vector literal
 * fix edge case in nested syntax-rules when variable in parent got expanded into identifier